requirements.txt
streamlit
Pillow
numpy
pycryptodome
